fixes integration tests
