+++ dkg-test-suite/package.json
@@ -4,11 +4,11 @@
 	"description": "e2e testing utilities for dkg-substrate",
 	"main": "index.js",
 	"scripts": {
-		"dkg-refresh": "ts-node ./src/evm/keyRefresh.ts",
+		"dkg-refresh": "ts-node ./src/keyRefresh.ts",
 		"anchor-proposals": "ts-node ./src/evm/testAnchorProposal.ts",
 		"anchor-create-proposal-substrate": "ts-node ./src/substrate/testAnchorCreateProposal.ts",
 		"anchor-update-proposal-substrate": "ts-node ./src/substrate/testAnchorUpdateProposal.ts",
-		"sudo-proposals": "ts-node ./src/evm/testDirectProposal.ts",
+		"sudo-proposals": "ts-node ./src/testDirectProposal.ts",
 		"max-deposit-limit-update-proposal": "ts-node ./src/evm/v-anchor-configurable-limit/testMaxDepositLimitUpdateProposal.ts",
 		"max-ext-limit-update-proposal": "ts-node ./src/evm/v-anchor-configurable-limit/testMaxExtLimitUpdate.ts",
 		"max-fee-limit-update-proposal": "ts-node ./src/evm/v-anchor-configurable-limit/testMaxFeeLimitUpdateProposal.ts",
