+++ dkg-test-suite/src/proposalTypes.ts
