+++ dkg-test-suite/src/substrate/testAnchorCreateProposal.ts
@@ -2,9 +2,11 @@ import {ApiPromise} from '@polkadot/api';
 import {Keyring} from '@polkadot/keyring';
 import {
 	encodeSubstrateProposal,
+} from './utils';
+import {
 	provider,
 	waitNfinalizedBlocks,
-} from './utils';
+} from '../utils';
 import {ethers} from 'ethers';
 import {keccak256} from '@ethersproject/keccak256';
 import {ECPair} from 'ecpair';
@@ -25,13 +27,13 @@ async function testAnchorCreateProposal() {
 		{compressed: false}
 	).publicKey.toString('hex');
 	const chainIdType = api.createType('DkgRuntimePrimitivesChainIdType', {SUBSTRATE: 5002});
-	const propHash = keccak256(encodeSubstrateProposal(anchorCreateProposal, 60));
+	const propHash = keccak256(encodeSubstrateProposal(anchorCreateProposal, 3000));
 
 	const proposalType = {anchorcreateproposal: anchorCreateProposal.header.nonce}
 
 	const unsubSignedProps: any = await unsubSignedPropsUtil(api, chainIdType, dkgPubKey, proposalType, propHash);
 
-	await new Promise((resolve) => setTimeout(resolve, 20000));
+	await new Promise((resolve) => setTimeout(resolve, 50000));
 
 	unsubSignedProps();
 }
@@ -45,7 +47,7 @@ async function sendAnchorCreateProposal(api: ApiPromise) {
 		api.query.dkg.dKGPublicKey(),
 	]);
 
-	const prop = u8aToHex(encodeSubstrateProposal(anchorCreateProposal, 60));
+	const prop = u8aToHex(encodeSubstrateProposal(anchorCreateProposal, 3000));
 	console.log(`DKG authority set id: ${authoritySetId}`);
 	console.log(`DKG pub key: ${dkgPubKey}`);
 	console.log(`Resource id is: ${resourceId}`);
