+++ dkg-test-suite/src/substrate/testWrappingFeeUpdateProposal.ts
@@ -2,9 +2,11 @@ import {ApiPromise} from '@polkadot/api';
 import {Keyring} from '@polkadot/keyring';
 import {
 	encodeSubstrateProposal,
+} from './utils';
+import {
 	provider,
 	waitNfinalizedBlocks,
-} from './utils';
+} from '../utils';
 import {keccak256} from '@ethersproject/keccak256';
 import {ECPair} from 'ecpair';
 import {assert, u8aToHex} from '@polkadot/util';
@@ -24,13 +26,13 @@ async function testWrappingFeeUpdateProposal() {
 		{compressed: false}
 	).publicKey.toString('hex');
 	const chainIdType = api.createType('DkgRuntimePrimitivesChainIdType', {SUBSTRATE: 5002});
-	const propHash = keccak256(encodeSubstrateProposal(wrappingFeeUpdateProposal, 40));
+	const propHash = keccak256(encodeSubstrateProposal(wrappingFeeUpdateProposal, 3000));
 
 	const proposalType = {wrappingfeeupdateproposal: wrappingFeeUpdateProposal.header.nonce}
 
 	const unsubSignedProps: any = await unsubSignedPropsUtil(api, chainIdType, dkgPubKey, proposalType, propHash);
 
-	await new Promise((resolve) => setTimeout(resolve, 20000));
+	await new Promise((resolve) => setTimeout(resolve, 50000));
 
 	unsubSignedProps();
 }
@@ -44,7 +46,7 @@ async function sendWrappingFeeUpdateProposal(api: ApiPromise) {
 		api.query.dkg.dKGPublicKey(),
 	]);
 
-	const prop = u8aToHex(encodeSubstrateProposal(wrappingFeeUpdateProposal, 40));
+	const prop = u8aToHex(encodeSubstrateProposal(wrappingFeeUpdateProposal, 3000));
 	console.log(`DKG authority set id: ${authoritySetId}`);
 	console.log(`DKG pub key: ${dkgPubKey}`);
 	console.log(`Resource id is: ${resourceId}`);
