+++ dkg-test-suite/src/utils.ts
@@ -1,0 +1,63 @@
+import { ApiPromise } from '@polkadot/api';
+import { WsProvider } from '@polkadot/api';
+
+export const ALICE = '5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY';
+
+export const provider = new WsProvider('ws://127.0.0.1:9944');
+
+export const hexToBytes = function (hex: any) {
+	for (var bytes = [], c = 0; c < hex.length; c += 2) {
+		bytes.push(parseInt(hex.substr(c, 2), 16));
+	}
+	return bytes;
+};
+
+export const listenOneBlock = async function (api: ApiPromise) {
+	const unsubscribe = await api.rpc.chain.subscribeNewHeads((header) => {
+		console.log(`Chain is at block: #${header.hash}`);
+		unsubscribe();
+	});
+};
+
+export const waitNfinalizedBlocks = async function (api: ApiPromise, n: number, timeout: number) {
+	return new Promise<void>(async (resolve, _reject) => {
+		let count = 0;
+		const unsubscribe = await api.rpc.chain.subscribeNewHeads((header) => {
+			count++;
+			if (count == n) {
+				unsubscribe();
+				resolve();
+			}
+			setTimeout(() => {
+				unsubscribe();
+				resolve();
+			}, timeout * 1000);
+		});
+	});
+};
+
+export const printValidators = async function (api: ApiPromise) {
+	const [{ nonce: accountNonce }, now, validators] = await Promise.all([
+		api.query.system.account(ALICE),
+		api.query.timestamp.now(),
+		api.query.session.validators(),
+	]);
+
+	console.log(`accountNonce(${ALICE}) ${accountNonce}`);
+	console.log(`last block timestamp ${now.toNumber()}`);
+
+	if (validators && validators.length > 0) {
+		const validatorBalances = await Promise.all(
+			validators.map((authorityId) => api.query.system.account(authorityId))
+		);
+
+		console.log(
+			'validators',
+			validators.map((authorityId, index) => ({
+				address: authorityId.toString(),
+				balance: validatorBalances[index].data.free.toHuman(),
+				nonce: validatorBalances[index].nonce.toHuman(),
+			}))
+		);
+	}
+};
